drivers: bme280: Add bme280_configure API
If a BME280 sensor experiences a reset, either due to a power cycle,
or spontaneously as has been observed by us, its config registers
need to be refreshed in order to restore normal operation.

This commit adds an API call to restore the configuration that
is otherwise only done on driver init.

Signed-off-by: Alexander Mihajlovic <alexander.mihajlovic@endian.se>