build: improve image signing
There is no mechanism to pass the image version during the
signing process.

To deal with it, a new variable called APPLICATION_IMG_VERSION is used
and has to be defined early on during the build process.

It has to be in a valid format for MCU boot.

The slot size is already known during built-time, so it makes sense
to pass the agument directly.

Signed-off-by: Efrain Calderon <efrain.calderon@aquarobur.com>